+++ packages/nimble-components/src/table-column/base/index.ts
@@ -1,5 +1,9 @@
-import type { ElementStyles, ViewTemplate } from '@microsoft/fast-element';
-import { attr } from '@microsoft/fast-element';
+import {
+    attr,
+    ElementStyles,
+    observable,
+    ViewTemplate
+} from '@microsoft/fast-element';
 import { FoundationElement } from '@microsoft/fast-foundation';
 import { uniqueId } from '@microsoft/fast-web-utilities';
 import type {
@@ -24,38 +28,55 @@ export abstract class TableColumn<
     @attr({ attribute: 'action-menu-label' })
     public actionMenuLabel?: string;
 
+    @attr({ attribute: 'column-hidden', mode: 'boolean' })
+    public columnHidden = false;
+
     /**
+     * @internal
+     *
      * The template to use to render the cell content for the column
      */
     // prettier-ignore
-    public abstract cellTemplate: ViewTemplate<TableCellState<TCellRecord, TColumnConfig>>;
+    public abstract readonly cellTemplate: ViewTemplate<TableCellState<TCellRecord, TColumnConfig>>;
 
     /**
+     * @internal
+     *
      * The style to apply to the cellTemplate
      */
-    public abstract cellStyles?: ElementStyles;
+    public abstract readonly cellStyles?: ElementStyles;
 
     /**
+     * @internal
+     *
      * The names of the fields that should be present in TCellRecord.
-     * This array is parallel with the field names returned from `getDataRecordFieldNames()`.
+     * This array is parallel with the field names specified by `dataRecordFieldNames`.
      */
     public abstract readonly cellRecordFieldNames: readonly TableFieldName[];
 
+    /**
+     * @internal
+     *
+     * The names of the fields from the row's record that correlate to the data that will be in TCellRecord.
+     * This array is parallel with the field names specified by `cellRecordFieldNames`.
+     */
+    @observable
+    public dataRecordFieldNames: readonly (TableFieldName | undefined)[] = [];
+
+    /**
+     * @internal
+     *
+     * The relevant, static configuration a column requires its cellTemplate to have access to.
+     */
+    @observable
+    public columnConfig?: TColumnConfig;
+
+    /**
+     * @internal
+     */
     public override connectedCallback(): void {
         super.connectedCallback();
 
         this.setAttribute('slot', uniqueId('table-column-slot'));
     }
-
-    /**
-     * This method returns the relevant, static configuration a column requires its cellTemplate
-     * to have access to.
-     */
-    public abstract getColumnConfig?(): TColumnConfig;
-
-    /**
-     * The names of the fields from the row's record that correlate to the data that will be in TCellRecord.
-     * This array is parallel with the field names specified by `cellRecordFieldNames`.
-     */
-    public abstract getDataRecordFieldNames(): (TableFieldName | undefined)[];
 }