+++ packages/nimble-components/src/table/template.ts
@@ -21,9 +21,11 @@ export const template = html<Table>`
             <div role="rowgroup" class="header-container">
                 <div class="header-row" role="row">
                     ${repeat(x => x.columns, html<TableColumn>`
-                        <${DesignSystem.tagFor(TableHeader)} class="header">
-                            <slot name="${x => x.slot}"></slot>
-                        </${DesignSystem.tagFor(TableHeader)}>
+                        ${when(x => !x.columnHidden, html<TableColumn>`
+                            <${DesignSystem.tagFor(TableHeader)} class="header">
+                                <slot name="${x => x.slot}"></slot>
+                            </${DesignSystem.tagFor(TableHeader)}>
+                        `)}
                     `)}
                     <div class="header-scrollbar-spacer" style="width: ${x => x.virtualizer.headerContainerMarginRight}px;"></div>
                 </div>